Implement delete pet action (only for authorized user)

+++ src/actions/actions.ts
@@ -41,9 +41,9 @@ export async function addPet(pet: unknown) {
 	await sleep(1000);
 
 	const session = await auth();
-		if (!session?.user) {
-			redirect("/login");
-		}
+	if (!session?.user) {
+		redirect("/login");
+	}
 
 	const validatedPet = petFormSchema.safeParse(pet);
 	if (!validatedPet.success) {
@@ -102,6 +102,13 @@ export async function editPet(petId: unknown, newPetData: unknown) {
 export async function checkoutPet(petId: unknown) {
 	await sleep(1000);
 
+	// authenticate check
+	const session = await auth();
+	if (!session?.user) {
+		redirect("/login");
+	}
+
+	// validation
 	const validatedPetId = petIdSchema.safeParse(petId);
 	if (!validatedPetId.success) {
 		return {
@@ -109,6 +116,25 @@ export async function checkoutPet(petId: unknown) {
 		};
 	}
 
+	// authorization check (users own pet)
+	const pet = await prisma.pet.findUnique({
+		where: {
+			id: validatedPetId.data,
+		},
+	});
+	if (!pet) {
+		return {
+			message: "Pet not found",
+		};
+	}
+
+	if (pet.userId !== session.user.id) {
+		return {
+			message: "Not authorized",
+		};
+	}
+
+	//database mutation
 	try {
 		await prisma.pet.delete({
 			where: {
